# Backend/java/src/main/resources/static/swagger.yaml
openapi: "3.0.3"
info:
  title: "NoteS API"
  version: "v 0.0.1"
servers:
  - url: "{scheme}://{domain}:{port}/{context}"
    description: Resource server
    variables:
      scheme:
        enum:
          - 'http'
          - 'https'
        default: 'http'
      domain:
        enum:
          - 'localhost'
          - 'notes.fvds.ru'
        default: 'localhost'
      port:
        enum:
          - '8080'
          - '9080'
          - '9180'
          - '5267'
          - '8443'
          - '80'
          - '443'
        default: '5267'
      context:
        enum:
          - ''
          - 'server'
        default: ''
components:
  schemas:
...
    AddTagRequestDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
      additionalProperties: false
...
    CreateTagRequestDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        color:
          maximum: 16777215
          minimum: 0
          type: integer
          format: int32
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
  securitySchemes:
    Bearer:
      type: oauth2
      description: API использует [OAuth2 authorizationCode](https://auth0.com/docs/get-started/authentication-and-authorization-flow/authorization-code-flow).
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/realms/NoteS/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/realms/NoteS/protocol/openid-connect/token
          scopes:
            read: Balea Server HTTP Api
security:
  - Bearer: [ ]
paths:
...
  "/api/public/{accountName}/notes/{pathNote}/tags":
    get:
      tags:
        - PublicTags
      description: Теги конкретной заметки
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
    post:
      tags:
        - PublicTags
      description: Добавить тег к заметке
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AddTagRequestDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/AddTagRequestDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/AddTagRequestDto"
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
  "/api/public/{accountName}/tags":
    post:
      tags:
        - PublicTags
      description: Создание тега
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateTagRequestDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/CreateTagRequestDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CreateTagRequestDto"
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
    get:
      tags:
        - PublicTags
      description: Список тегов пользователя
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
  "/api/public/{accountName}/notes/{pathNote}/tags/{tagName}":
    delete:
      tags:
        - PublicTags
      description: Удаление тега для заметки
      parameters:
        - name: accountName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
        - name: pathNote
          in: path
          required: true
          schema:
            maxLength: 128
            minLength: 3
            type: string
        - name: tagName
          in: path
          required: true
          schema:
            maxLength: 128
            type: string
      responses:
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            application/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"
            text/json:
              schema:
                "$ref": "#/components/schemas/ProblemDetails"